requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
firebase-admin==6.7.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.36.0
db-dtypes==1.4.2
google-auth-oauthlib==1.0.0
oauthlib==3.2.2
pypdf==5.4.0
google-cloud-bigquery==3.31.0
google-cloud-pubsub==2.29.0
google-generativeai==0.8.4
python-multipart==0.0.20